+++ CHANGELOG.rst
@@ -49,6 +49,8 @@ Changed
   Contributed by Nick Maludy.
 * Update various Python dependencies to the latest stable versions (kombu, amqp, apscheduler,
   gitpython, pymongo, stevedore, paramiko, prompt-toolkit, flex). #3830
+* Update log messages in the datastore service to correctly use ``DEBUG`` log level instead of
+  ``AUDIT``. #3845
 
 Fixed
 ~~~~~